Modifications to the build CMakeLists to do versioning on the shared library as well as the following   - Fixed a warning regarding the response code.   - Added Release and Debug flags

+++ src/pillowtalk.h
@@ -21,7 +21,7 @@ typedef struct {
 
 typedef struct {
   pt_node_t* root;
-  int response_code;
+  long response_code;
 } pt_response_t;
 
 void pillowtalk_init();
